halo2_gadgets::ecc::chip

Module constants

Constants required for the ECC chip.

Constants

FIXED_BASE_WINDOW_SIZE

Window size for fixed-base scalar multiplication

H

$2^{FIXED_BASE_WINDOW_SIZE}$

NUM_WINDOWS

Number of windows for a full-width scalar

NUM_WINDOWS_SHORT

Number of windows for a short signed scalar

Functions

compute_lagrange_coeffs

For each window, we interpolate the $x$-coordinate. Here, we pre-compute and store the coefficients of the interpolation polynomial.

find_zs_and_us

For each window, $z$ is a field element such that for each point $(x, y)$ in the window:

test_lagrange_coeffstest-dependencies

Test that Lagrange interpolation coefficients reproduce the correct x-coordinate for each fixed-base multiple in each window.

test_zs_and_ustest-dependencies

Test that the z-values and u-values satisfy the conditions: 1. z + y = u^2, 2. z - y is not a square for the y-coordinate of each fixed-base multiple in each window.