Trait halo2_gadgets::ecc::chip::ScalarKind

pub trait ScalarKind {
    const NUM_WINDOWS: usize;
}

A trait representing the kind of scalar used with a particular FixedPoint.

This trait exists because of limitations around const generics.

Required Associated Constants

const NUM_WINDOWS: usize

The number of windows that this scalar kind requires.