Trait halo2_gadgets::ecc::BaseFitsInScalarInstructions

pub trait BaseFitsInScalarInstructions<C: CurveAffine>: EccInstructions<C> {
    fn scalar_var_from_base(
        &self,
        layouter: &mut impl Layouter<C::Base>,
        base: &Self::Var
    ) -> Result<Self::ScalarVar, Error>;
}

Instructions that can be implemented for a curve whose base field fits into its scalar field.

Required Methods

fn scalar_var_from_base(
&self,
layouter: &mut impl Layouter<C::Base>,
base: &Self::Var
) -> Result<Self::ScalarVar, Error>

Converts a base field element that exists as a variable in the circuit into a scalar to be used in variable-base scalar multiplication.