halo2_gadgets

Module poseidon

The Poseidon algebraic hash function.

Re-exports

pub use ::halo2_poseidon as primitives;

Structs

Hash

A Poseidon hash function, built around a sponge.

Pow5Chip

A Poseidon chip using an $x^5$ S-Box.

Pow5Config

Configuration for a Pow5Chip.

Sponge

A Poseidon sponge.

StateWord

A word in the Poseidon state.

Word

A word over which the Poseidon permutation operates.

Enums

PaddedWord

A word from the padded input to a Poseidon sponge.

Traits

PoseidonInstructions

The set of circuit instructions required to use the Poseidon permutation.

PoseidonSpongeInstructions

The set of circuit instructions required to use the Sponge and Hash gadgets.