Module halo2_gadgets::poseidon::primitives

The Poseidon algebraic hash function.

Structs

Absorbing

The absorbing state of the Sponge.

ConstantLength

A Poseidon hash function used with constant input length.

Hash

A Poseidon hash function, built around a sponge.

P128Pow5T3

Poseidon-128 using the $x^5$ S-box, with a width of 3 field elements, and the standard number of rounds for 128-bit security “with margin”.

Squeezing

The squeezing state of the Sponge.

Traits

Domain

A domain in which a Poseidon hash function is being used.

Spec

A specification for a Poseidon permutation.

SpongeMode

The state of the Sponge.

Functions

generate_constants

Generates (round_constants, mds, mds^-1) corresponding to this specification.

Type Aliases

Mds

The type used to hold the MDS matrix and its inverse.