halo2_gadgets::poseidon

Trait PoseidonInstructions

pub trait PoseidonInstructions<F: Field, S: Spec<F, T, RATE>, const T: usize, const RATE: usize>: Chip<F> {
    type Word: Clone + Debug + From<AssignedCell<F, F>> + Into<AssignedCell<F, F>>;

    // Required method
    fn permute(
        &self,
        layouter: &mut impl Layouter<F>,
        initial_state: &State<Self::Word, T>,
    ) -> Result<State<Self::Word, T>, Error>;
}

The set of circuit instructions required to use the Poseidon permutation.

Required Associated Types

type Word: Clone + Debug + From<AssignedCell<F, F>> + Into<AssignedCell<F, F>>

Variable representing the word over which the Poseidon permutation operates.

Required Methods

fn permute( &self, layouter: &mut impl Layouter<F>, initial_state: &State<Self::Word, T>, ) -> Result<State<Self::Word, T>, Error>

Applies the Poseidon permutation to the given state.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.