halo2_gadgets::poseidon

Trait PoseidonSpongeInstructions

pub trait PoseidonSpongeInstructions<F: Field, S: Spec<F, T, RATE>, D: Domain<F, RATE>, const T: usize, const RATE: usize>: PoseidonInstructions<F, S, T, RATE> {
    // Required methods
    fn initial_state(
        &self,
        layouter: &mut impl Layouter<F>,
    ) -> Result<State<Self::Word, T>, Error>;
    fn add_input(
        &self,
        layouter: &mut impl Layouter<F>,
        initial_state: &State<Self::Word, T>,
        input: &Absorbing<PaddedWord<F>, RATE>,
    ) -> Result<State<Self::Word, T>, Error>;
    fn get_output(state: &State<Self::Word, T>) -> Squeezing<Self::Word, RATE>;
}

The set of circuit instructions required to use the Sponge and Hash gadgets.

Required Methods

fn initial_state( &self, layouter: &mut impl Layouter<F>, ) -> Result<State<Self::Word, T>, Error>

Returns the initial empty state for the given domain.

fn add_input( &self, layouter: &mut impl Layouter<F>, initial_state: &State<Self::Word, T>, input: &Absorbing<PaddedWord<F>, RATE>, ) -> Result<State<Self::Word, T>, Error>

Adds the given input to the state.

fn get_output(state: &State<Self::Word, T>) -> Squeezing<Self::Word, RATE>

Extracts sponge output from the given state.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.