halo2_gadgets::sinsemilla

Module merkle

Gadgets for implementing a Merkle tree with Sinsemilla.

Modules

chip

Chip implementing a Merkle hash using Sinsemilla as the hash function.

Structs

MerklePath

Gadget representing a Merkle path that proves a leaf exists in a Merkle tree at a specific position.

Constants

MERKLE_CRH_PERSONALIZATION

SWU hash-to-curve personalization for the Merkle CRH generator

Traits

MerkleInstructions

Instructions to check the validity of a Merkle path of a given PATH_LENGTH. The hash function used is a Sinsemilla instance with K-bit words. The hash function can process MAX_WORDS words.