halo2_gadgets::utilities::cond_swap

Trait CondSwapInstructions

pub trait CondSwapInstructions<F: Field>: UtilitiesInstructions<F> {
    // Required methods
    fn swap(
        &self,
        layouter: impl Layouter<F>,
        pair: (Self::Var, Value<F>),
        swap: Value<bool>,
    ) -> Result<(Self::Var, Self::Var), Error>;
    fn mux(
        &self,
        layouter: &mut impl Layouter<F>,
        choice: Self::Var,
        left: Self::Var,
        right: Self::Var,
    ) -> Result<Self::Var, Error>;
}

Instructions for conditional swap and mux gadgets.

Required Methods

fn swap( &self, layouter: impl Layouter<F>, pair: (Self::Var, Value<F>), swap: Value<bool>, ) -> Result<(Self::Var, Self::Var), Error>

Given an input pair (a,b) and a swap boolean flag, returns (b,a) if swap is set, else (a,b) if swap is not set.

The second element of the pair is required to be a witnessed value, not a variable that already exists in the circuit.

fn mux( &self, layouter: &mut impl Layouter<F>, choice: Self::Var, left: Self::Var, right: Self::Var, ) -> Result<Self::Var, Error>

Given an input (choice, left, right) where choice is a boolean flag, returns left if choice is not set and right if choice is set.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.