halo2_gadgets::utilities

Function decompose_word

pub fn decompose_word<F: PrimeFieldBits>(
    word: &F,
    word_num_bits: usize,
    window_num_bits: usize,
) -> Vec<u8> 

Decompose a word alpha into window_num_bits bits (little-endian) For a window size of w, this returns [k_0, …, k_n] where each k_i is a w-bit value, and scalar = k_0 + k_1 * w + k_n * w^n.

Panics

We are returning a Vec<u8> which means the window size is limited to <= 8 bits.