halo2_gadgets::utilities::lookup_range_check

Trait LookupRangeCheck

pub trait LookupRangeCheck<F: PrimeFieldBits, const K: usize>:
    Eq
    + Copy
    + Debug {
    // Required methods
    fn config(&self) -> &LookupRangeCheckConfig<F, K>;
    fn configure(
        meta: &mut ConstraintSystem<F>,
        running_sum: Column<Advice>,
        table_idx: TableColumn,
    ) -> Self
       where Self: Sized;
    fn load(
        &self,
        generator_table_config: &GeneratorTableConfig,
        layouter: &mut impl Layouter<Base>,
    ) -> Result<(), Error>;
    fn short_range_check(
        &self,
        region: &mut Region<'_, F>,
        element: AssignedCell<F, F>,
        num_bits: usize,
    ) -> Result<(), Error>;

    // Provided methods
    fn copy_check(
        &self,
        layouter: impl Layouter<F>,
        element: AssignedCell<F, F>,
        num_words: usize,
        strict: bool,
    ) -> Result<RunningSum<F>, Error> { ... }
    fn witness_check(
        &self,
        layouter: impl Layouter<F>,
        value: Value<F>,
        num_words: usize,
        strict: bool,
    ) -> Result<RunningSum<F>, Error> { ... }
    fn range_check(
        &self,
        region: &mut Region<'_, F>,
        element: AssignedCell<F, F>,
        num_words: usize,
        strict: bool,
    ) -> Result<RunningSum<F>, Error> { ... }
    fn copy_short_check(
        &self,
        layouter: impl Layouter<F>,
        element: AssignedCell<F, F>,
        num_bits: usize,
    ) -> Result<(), Error> { ... }
    fn witness_short_check(
        &self,
        layouter: impl Layouter<F>,
        element: Value<F>,
        num_bits: usize,
    ) -> Result<AssignedCell<F, F>, Error> { ... }
}

Trait that provides common methods for a lookup range check.

Required Methods

fn config(&self) -> &LookupRangeCheckConfig<F, K>

Returns a reference to the LookupRangeCheckConfig instance.

fn configure( meta: &mut ConstraintSystem<F>, running_sum: Column<Advice>, table_idx: TableColumn, ) -> Self
where Self: Sized,

The running_sum advice column breaks the field element into K-bit words. It is used to construct the input expression to the lookup argument.

The table_idx fixed column contains values from [0..2^K). Looking up a value in table_idx constrains it to be within this range. The table can be loaded outside this helper.

Side-effects

Both the running_sum and constants columns will be equality-enabled.

fn load( &self, generator_table_config: &GeneratorTableConfig, layouter: &mut impl Layouter<Base>, ) -> Result<(), Error>

Load the generator table into the circuit.

fn short_range_check( &self, region: &mut Region<'_, F>, element: AssignedCell<F, F>, num_bits: usize, ) -> Result<(), Error>

Constrain x to be a NUM_BITS word.

element must have been assigned to self.config().running_sum at offset 0.

Provided Methods

fn copy_check( &self, layouter: impl Layouter<F>, element: AssignedCell<F, F>, num_words: usize, strict: bool, ) -> Result<RunningSum<F>, Error>

Range check on an existing cell that is copied into this helper.

Returns an error if element is not in a column that was passed to ConstraintSystem::enable_equality during circuit configuration.

fn witness_check( &self, layouter: impl Layouter<F>, value: Value<F>, num_words: usize, strict: bool, ) -> Result<RunningSum<F>, Error>

Range check on a value that is witnessed in this helper.

fn range_check( &self, region: &mut Region<'_, F>, element: AssignedCell<F, F>, num_words: usize, strict: bool, ) -> Result<RunningSum<F>, Error>

If strict is set to “true”, the field element must fit into num_words * K bits. In other words, the the final cumulative sum z_{num_words} must be zero.

If strict is set to “false”, the final z_{num_words} is not constrained.

element must have been assigned to self.running_sum at offset 0.

fn copy_short_check( &self, layouter: impl Layouter<F>, element: AssignedCell<F, F>, num_bits: usize, ) -> Result<(), Error>

Short range check on an existing cell that is copied into this helper.

Panics

Panics if NUM_BITS is equal to or larger than K.

fn witness_short_check( &self, layouter: impl Layouter<F>, element: Value<F>, num_bits: usize, ) -> Result<AssignedCell<F, F>, Error>

Short range check on value that is witnessed in this helper.

Panics

Panics if num_bits is larger than K.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.