halo2_gadgets::utilities::lookup_range_check

Type Alias PallasLookupRangeCheck4_5BConfig

pub type PallasLookupRangeCheck4_5BConfig = LookupRangeCheck4_5BConfig<Base, { sinsemilla::K }>;

In this crate, LookupRangeCheck4_5BConfig is always used with pallas::Base as the prime field and the constant K from the sinsemilla module. To reduce verbosity and improve readability, we introduce this alias and use it instead of that long construction.

Aliased Type

pub struct PallasLookupRangeCheck4_5BConfig { /* private fields */ }