halo2_gadgets

Module ecc

Elliptic curve operations.

Modules

chip

Chip implementations for the ECC gadgets.

Structs

FixedPoint

Precomputed multiples of a fixed point, for full-width scalar multiplication.

FixedPointBaseField

Precomputed multiples of a fixed point, that can be multiplied by base-field elements.

FixedPointShort

Precomputed multiples of a fixed point, for short signed scalar multiplication.

NonIdentityPoint

A point on a specific elliptic curve that is guaranteed to not be the identity.

Point

A point on a specific elliptic curve.

ScalarFixed

An integer representing an element of the scalar field for a specific elliptic curve, for FixedPoint scalar multiplication.

ScalarFixedShort

A signed short (64-bit) integer represented as an element of the scalar field for a specific elliptic curve, to be used for FixedPointShort scalar multiplication.

ScalarVar

An integer representing an element of the scalar field for a specific elliptic curve.

X

The affine short Weierstrass x-coordinate of a point on a specific elliptic curve.

Traits

BaseFitsInScalarInstructions

Instructions that can be implemented for a curve whose base field fits into its scalar field.

EccInstructions

The set of circuit instructions required to use the ECC gadgets.

FixedPoints

Defines the fixed points for a given instantiation of the ECC chip.