Module halo2_gadgets::ecc

Gadgets for elliptic curve operations.

Modules

chip

Chip implementations for the ECC gadgets.

Structs

FixedPoint

A constant elliptic curve point over the given curve, for which window tables have been provided to make scalar multiplication more efficient.

FixedPointBaseField

A constant elliptic curve point over the given curve, used in scalar multiplication with a base field element

FixedPointShort

A constant elliptic curve point over the given curve, used in scalar multiplication with a short signed exponent

NonIdentityPoint

A non-identity elliptic curve point over the given curve.

Point

An elliptic curve point over the given curve.

ScalarFixed

A full-width element of the given elliptic curve’s scalar field, to be used for fixed-base scalar mul.

ScalarFixedShort

A signed short element of the given elliptic curve’s scalar field, to be used for fixed-base scalar mul.

ScalarVar

An element of the given elliptic curve’s base field, that is used as a scalar in variable-base scalar mul.

X

The affine short Weierstrass x-coordinate of an elliptic curve point over the given curve.

Traits

EccInstructions

The set of circuit instructions required to use the ECC gadgets.

FixedPoints

Defines the fixed points for a given instantiation of the ECC chip.