Function halo2_gadgets::ecc::chip::constants::find_zs_and_us

pub fn find_zs_and_us<C: CurveAffine>(
    base: C,
    num_windows: usize
) -> Option<Vec<(u64, [C::Base; 8])>>

For each window, $z$ is a field element such that for each point $(x, y)$ in the window:

$z + y = u^2$ (some square in the field); and

$z - y$ is not a square. If successful, return a vector of (z: u64, us: [C::Base; H]) for each window.

This function was used to generate the zs and us for the Orchard fixed bases. The outputs of this function have been stored as constants, and it is not called anywhere in this codebase. However, we keep this function here as a utility for those who wish to use it with different parameters.