Trait halo2_gadgets::poseidon::primitives::Domain

pub trait Domain<F: Field, const RATE: usize> {
    type Padding: IntoIterator<Item = F>;

    // Required methods
    fn name() -> String;
    fn initial_capacity_element() -> F;
    fn padding(input_len: usize) -> Self::Padding;
}

A domain in which a Poseidon hash function is being used.

Required Associated Types

type Padding: IntoIterator<Item = F>

Iterator that outputs padding field elements.

Required Methods

fn name() -> String

The name of this domain, for debug formatting purposes.

fn initial_capacity_element() -> F

The initial capacity element, encoding this domain.

fn padding(input_len: usize) -> Self::Padding

Returns the padding to be appended to the input.