Module halo2_gadgets::sinsemilla

The Sinsemilla hash function.

Modules

chip

Chip implementations for the Sinsemilla gadgets.

merkle

Gadgets for implementing a Merkle tree with Sinsemilla.

primitives

Implementation of Sinsemilla outside the circuit.

Structs

CommitDomain

Gadget representing a domain in which $\mathsf{SinsemillaCommit}$ and $\mathsf{SinsemillaShortCommit}$ can be used.

HashDomain

A domain in which $\mathsf{SinsemillaHashToPoint}$ and $\mathsf{SinsemillaHash}$ can be used.

Message

A message to be hashed.

MessagePiece

A message piece with a bitlength of some multiple of K.

Traits

CommitDomains

Trait allowing circuit’s Sinsemilla CommitDomains to be enumerated.

HashDomains

Trait allowing circuit’s Sinsemilla HashDomains to be enumerated.

SinsemillaInstructions

The set of circuit instructions required to use the Sinsemilla gadget. This trait is bounded on two constant parameters: K, the number of bits in each word accepted by the Sinsemilla hash, and MAX_WORDS, the maximum number of words that a single hash instance can process.